e2e: refactor grpc stats test
