+++ tests/integration/telemetry/envoyfilter/customizemetrics/customize_metrics_test.go
@@ -151,41 +151,6 @@ func TestCustomizeMetrics(t *testing.T) {
 		})
 }
 
-// TestCustomGRPCMetrics tests that istio_[request/response]_messages_total are present https://github.com/istio/istio/issues/44144
-// Kiali depends on these metrics
-func TestCustomGRPCCountMetrics(t *testing.T) {
-	framework.NewTest(t).
-		Label(label.IPv4). // https://github.com/istio/istio/issues/35835
-		Features("observability.telemetry.stats.prometheus.customize-metric").
-		Run(func(t framework.TestContext) {
-			// Metrics to be queried and tested
-			metrics := [2]string{"istio_request_messages_total", "istio_response_messages_total"}
-			for _, metric := range metrics {
-				t.NewSubTestf(metric).Run(func(t framework.TestContext) {
-					t.Cleanup(func() {
-						if t.Failed() {
-							util.PromDump(t.Clusters().Default(), promInst, prometheus.Query{Metric: metric})
-						}
-						grpcSourceQuery := buildGRPCQuery(metric)
-						cluster := t.Clusters().Default()
-						retry.UntilSuccessOrFail(t, func() error {
-							if err := sendTraffic(); err != nil {
-								t.Log("failed to send grpc traffic")
-								return err
-							}
-							if _, err := util.QueryPrometheus(t, cluster, grpcSourceQuery, promInst); err != nil {
-								util.PromDiff(t, promInst, cluster, grpcSourceQuery)
-								return err
-							}
-							return nil
-						}, retry.Delay(1*time.Second), retry.Timeout(300*time.Second))
-						util.ValidateMetric(t, cluster, promInst, grpcSourceQuery, 1)
-					})
-				})
-			}
-		})
-}
-
 func TestMain(m *testing.M) {
 	framework.NewSuite(m).
 		Label(label.CustomSetup).
@@ -400,22 +365,6 @@ func buildQuery(protocol string) (destinationQuery prometheus.Query) {
 
 	_, destinationQuery, _ = common.BuildQueryCommon(labels, appNsInst.Name())
 	return destinationQuery
-}
-
-func buildGRPCQuery(metric string) (destinationQuery prometheus.Query) {
-	labels := map[string]string{
-		"destination_app":                "server",
-		"destination_version":            "v1",
-		"destination_service":            "server." + appNsInst.Name() + ".svc.cluster.local",
-		"destination_service_name":       "server",
-		"destination_workload_namespace": appNsInst.Name(),
-		"destination_service_namespace":  appNsInst.Name(),
-	}
-	sourceQuery := prometheus.Query{}
-	sourceQuery.Metric = metric
-	sourceQuery.Labels = labels
-
-	return sourceQuery
 }
 
 func createDockerCredential(user, passwd, registry string) string {
