+++ tests/integration/telemetry/envoyfilter/nullvm/stats_filter_test.go
@@ -36,6 +36,10 @@ func TestStatsFilter(t *testing.T) {
 	common.TestStatsFilter(t, features.Feature("observability.telemetry.stats.prometheus.http.nullvm"), common.DefaultBucketCount)
 }
 
+func TestGRPCStats(t *testing.T) {
+	common.TestGRPCMetrics(t, features.Feature("observability.telemetry.stats.prometheus.grpc.nullvm"))
+}
+
 func TestMain(m *testing.M) {
 	framework.NewSuite(m).
 		Label(label.CustomSetup).
